Backend/requirement.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==5.0.0
attrs==25.1.0
billiard==4.2.1
bitarray==3.1.0
blinker==1.8.2
CacheControl==0.14.3
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
cytoolz==1.0.1
dataclasses==0.6
Django==5.1.1
django-cors-headers==4.5.0
django-environ==0.12.0
django-filter==24.2
django-telegram-login==0.2.3
django_paystack==1.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.2.31
dnspython==2.7.0
dotenv==0.9.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.0.1
exceptiongroup==1.2.2
firebase-admin==6.8.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
forex-python==1.8
frozenlist==1.5.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==4.0.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.4.2
lru-dict==1.3.0
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.1.0
packaging==24.1
parsimonious==0.9.0
paystackapi==2.1.3
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.3
pypaystack2==2.0.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.6
python_paystack==1.2.0
pyunormalize==16.0.0
redis==5.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.24.0
rsa==4.9.1
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
toolz==1.0.0
tornado==6.4.1
types-requests==2.32.0.20250301
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.1
validators==0.34.0
vine==5.1.0
wcwidth==0.2.13
web3==6.0.0
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
yarl==1.18.3
